requirements.txt
certifi==2019.6.16
chardet==3.0.4
cssselect==1.1.0
Django==2.2.5
django-cors-headers==3.1.0
django-extensions==2.2.1
djangorestframework==3.10.3
idna==2.8
lxml==4.4.1
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
